v1.0.4

+++ src/main/java/com/potato/domain/UserVO.java
@@ -8,7 +8,7 @@ public class UserVO {
 	//UserVO
 		private String user_number;	  //유저번호(=회원번호) sys_guid()
 		private int reports;		  //신고당한 수
-		private int temper;			  //온도
+		private Long temper;			  //온도
 		private int trades;			  //거래완료 수
 
 }
